requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.1.0
scipy>=1.9.0

# Deep Learning
torch>=1.12.0
sentence-transformers>=2.2.0

# Clustering
hdbscan>=0.8.29

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web App
streamlit>=1.12.0

# Text Processing
nltk>=3.7
spacy>=3.4.0

# Testing
pytest>=7.0.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.64.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.15.0
